add extra state. rename. clean up states.
