+++ lib/components/proposals/ProposalsUI.jsx
@@ -5,7 +5,7 @@ import { AddGovernanceTokenToMetaMask } from 'lib/components/AddGovernanceTokenT
 import { Button } from 'lib/components/Button'
 import { ProposalsList } from 'lib/components/proposals/ProposalsList'
 import { RetroactivePoolClaimBanner } from 'lib/components/RetroactivePoolClaimBanner'
-import { PoolVotesCard, UsersVotesCard } from 'lib/components/UsersVotesCard'
+import { UsersPoolVotesCard } from 'lib/components/UsersPoolVotesCard'
 import { V3LoadingDots } from 'lib/components/V3LoadingDots'
 import { useAllProposalsSorted } from 'lib/hooks/useAllProposalsSorted'
 import { ButtonLink } from 'lib/components/ButtonLink'
@@ -31,9 +31,7 @@ export const ProposalsUI = (props) => {
     <>
       <RetroactivePoolClaimBanner />
 
-      <UsersVotesCard />
-
-      <PoolVotesCard />
+      <UsersPoolVotesCard />
 
       <h1 className='text-inverse'>{t('vote')}</h1>
