+++ lib/hooks/useTokenHolder.js
@@ -63,8 +63,6 @@ async function getTokenHolder (address, chainId, blockNumber) {
       variables
     )
 
-    console.log('subgraphData', subgraphData)
-
     if (!subgraphData.tokenHolder && !subgraphData.delegate) {
       return {
         tokenBalance: '0',
